Add `UrlTarget` to `New-HTMLCalendar` allowing for greater link control.
The `New-HTMLCalendar` function now accepts an additional parameter called `UrlTarget`, which allows for greater control over links within the calendar. This parameter can be used to specify the target for the URL, such as opening the link in an iframe. The code has been updated to handle this new parameter and modify the generated calendar accordingly.